refactor: 设置页和实验功能水印
