+++ front_end/src/i18n/locales/zh-cn.ts
@@ -200,6 +200,7 @@ export default {
             panic: '录像解析器发生未知错误，请联系开发者',
         },
     },
+    experimentalFeature: '实验功能',
     form: {
         captcha: '验证码',
         captchaLoading: '加载中',
